WIP: Remove installation of Authenticators from pgm.Main
When pushing with JGit commandline to e.g. Github exceptions occur
without this fix.

The problem is that during a push two requests are sent to the server,
first a GET and then a POST (containing the pack data). The first GET
request sent anonymously is rejected with 401 (unauthorized). When an
Authenticator is installed the java.net classes will use the
Authenticator to ask the user for credentials and retry the request.
But this happens under the hood and JGit level code doesn't see that
this happens.
The next request is the POST but since JGit thinks the first GET request
went through anonymously it doesn't add authentication headers to the
POST request. This POST of course also fails with 401 but since this
request contains a lot of body-data streamed from JGit (the pack file!)
the java.net classes can't simply retry the request with authorization
headers. The whole process fails.

If no Authenticator is installed then the situation is better. First
anonymous GET request fails with 401 and this bubbles up to JGit. Then
JGit asks the user for credentials (with the help of the installed
CredentialsProviders) and retries with basic authentication headers.
This succeeds and when the POST request is sent JGit knows about the
succesfull authorization and adds the correct header directly to the
POST. The request succeeds.

In this proposal I simply remove the authenticator from JGit pgm.
Anybody knows what will break now?

Change-Id: Ic50a0685cdf5b1aef77426c115f19ef3f4fab156